# [XPU] IPEX-optimized Punica Wrapper on XPU