--- a/vllm/lora/punica_wrapper/punica_xpu.py
+++ b/vllm/lora/punica_wrapper/punica_xpu.py
@@ -22,7 +22,7 @@ class PunicaWrapperXPU(PunicaWrapperBase):
     """
     PunicaWrapperXPU is designed to manage and provide metadata for the punica
     kernel. The main function is to maintain the state information for 
-    Multi-LoRA, and to provide the interface for the punica triton kernel.
+    Multi-LoRA, and to provide the interface for the punica ipex kernel.
     """
 
     def __init__(self, max_num_batched_tokens: int, max_batches: int,